third_party/blink/web_tests/fast/forms/color/color-picker-show-picker.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script>
  promise_test(async (t) => {
    const input = document.createElement("input");
    input.setAttribute("type", "color");

    await test_driver.bless("show color picker");
    input.showPicker();
    assert_not_equals(internals.pagePopupWindow, null, "Picker is opened.");

    internals.endColorChooser(input);
    assert_equals(internals.pagePopupWindow, null, "Picker is closed.");
  }, "Color picker: showPicker() should show picker");
</script>
